--- src/_pages/_app.tsx
@@ -1,23 +1,0 @@
-import { type AppType } from 'next/app';
-import { type Session } from 'next-auth';
-import { SessionProvider } from 'next-auth/react';
-
-import { api } from '@src/utils/api';
-import '@src/styles/globals.css';
-import Sidebar from '@src/components/Sidebar';
-
-const MyApp: AppType<{ session: Session | null }> = ({
-  Component,
-  pageProps: { session, ...pageProps },
-}) => {
-  return (
-    <SessionProvider {...session}>
-      <Sidebar />
-      <div className="max-h-screen overflow-y-scroll">
-        <Component {...pageProps} />
-      </div>
-    </SessionProvider>
-  );
-};
-
-export default api.withTRPC(MyApp);
