--- src/_pages/api/auth/[...nextauth].ts
@@ -1,4 +1,0 @@
-import NextAuth from 'next-auth';
-import { authOptions } from '@src/server/auth';
-
-export default NextAuth(authOptions);
