+++ src/components/OrgHeader.tsx
@@ -5,13 +5,11 @@ import type {
   SelectContact as Contacts,
 } from '@src/server/db/models';
 import { HeartIcon } from './Icons';
-import { useState } from 'react';
 
 type Club = SelectClub & {
   contacts?: Contacts[];
 };
 const OrgHeader = ({ club }: { club: Club }) => {
-  const [liked, setLiked] = useState(false);
   return (
     <div className="relative">
       <div className="h-full w-full">
@@ -48,12 +46,9 @@ const OrgHeader = ({ club }: { club: Club }) => {
             <button
               className="rounded-full bg-blue-primary p-2.5 transition-colors hover:bg-blue-700"
               type="button"
-              onClick={() => {
-                setLiked(!liked);
-              }}
             >
               <div className={'h-8 w-8'}>
-                <HeartIcon fill={`${liked ? 'fill-red-500' : 'fill-white'} `} />
+                <HeartIcon fill="fill-white" />
               </div>
             </button>
             <ContactButtons contacts={club.contacts || []} />
